chore: 모델별 파일 분리
